Cell viz: modularize + remove failed song wave
Split the single ~1270-line CellVisualization.kt into three files with
clear dependency boundaries:

  CellVizConfig.kt (216 lines)  — public config data class + DEFAULT
  CellVizModel.kt  (289 lines)  — internal pure domain/helpers (no Compose)
  CellVisualization.kt (~550 lines) — Compose-dependent: composable + draws

Every helper in CellVizModel.kt is `internal` so commonTest can reach
in without making the API public. The file has zero Compose imports,
so tests run fast and don't need a Compose test harness — pure kotlin.test
assertions against math-only functions.

Motivations for the split:
  1. The file was getting long. 1270 lines is past where a reviewer
     stops holding the whole thing in head.
  2. Tests. The non-Compose helpers (MembraneOpening lifecycle,
     subtractRangesFromArc, openingRanges wrapping, smoothstep,
     adapterBus, spreadAngle, spawnOpening, CellVizConfig.sanitized)
     are the most test-worthy code in the viz — they are all pure math
     with no UI dependencies. Splitting them into a Compose-free file
     makes adding tests a trivial next step (in flight).
  3. Dependency clarity. CellVizConfig depends on nothing. CellVizModel
     depends on CellVizConfig + kotlin.math + kotlin.random. Only
     CellVisualization.kt depends on Compose. Clean pyramid.

Dropped along with the split:
  - The nucleus "song" wave, after multiple tuning attempts failed to
    make it read as moving against the mote cloud. See the retained
    comment block in CellVisualization.kt for the full lesson.
  - Song-related config fields: nucleusSongPeriodSec,
    nucleusSongPeakOpacity, nucleusSongEmissionEveryN. Breathing +
    nucleus radius remain.

Public API surface is unchanged: `CellVisualization(...)` composable
and `CellVizConfig(...)` data class stay the same. Only internal
code moved.
