+++ mobile/shared/src/commonMain/kotlin/ai/ciris/mobile/shared/ui/screens/graph/CellVisualization.kt
@@ -25,301 +25,35 @@ import kotlin.math.cos
 import kotlin.math.sin
 
 // =============================================================================
-// Cell Visualization — Step 3: membrane skeleton + seam + static adapter ports
+// CellVisualization — Compose-dependent rendering of the cell
 // =============================================================================
 //
-// This composable is the new Interact-screen background. It replaces the
-// cylinder-based [LiveGraphBackground] on devices that pass the cell-viz
-// capability gate (64-bit ABI, sufficient RAM). See FSD/CELL_VIZ_REDESIGN.md
-// for the full design document.
+// Pure domain model + math lives in [CellVizModel.kt]. User-tunable config
+// lives in [CellVizConfig.kt]. This file is the only one that depends on
+// Compose, so it stays small-ish and every non-draw function we own can be
+// unit-tested from commonTest against the model file.
 //
-// What this step renders (static, no events yet):
+// See FSD/CELL_VIZ_REDESIGN.md for the full design. Quick map of which
+// steps are wired in here:
 //
-//  - The medium (dark radial-gradient background; light mode is a parity
-//    fallback, not the composition target).
-//  - The cell-body aura (faint warm wash inside the membrane).
-//  - Six bus arcs forming the membrane, each with a three-path bloom
-//    stack — outer halo, mid halo, bright stroke. Colors per §2 of the
-//    design doc: wise/runtime/tool/llm/memory/comm at fixed angles.
-//  - The membrane seam — a single 6° gap inside the Wise arc. Visible
-//    acknowledgement that the system is incomplete by design.
-//  - Adapter ports — diamonds or hexagons anchored on their owning bus
-//    arc, spread evenly within the 60° segment. Shape and color come
-//    from the adapter type.
-//
-// What this step does NOT render (intentional — later steps):
-//
-//  - The nucleus (step 4)
-//  - Cytoplasm motes (step 5)
-//  - Bus shimmer / gratitude motes / event pulses (steps 6, 8)
-//  - Pseudopods + weaving tendrils (step 7)
-//  - Deferral ripple + WA companion cell (step 9)
-//  - BG↔FG zoom transition (step 10)
-//
-// Rotation matches the step-2 refactor: a single withFrameNanos-driven
-// variable, 6°/sec baseline, +external swipe rotation.
-//
-// CONFIGURATION:
-//
-// Every tunable the viz exposes lives on [CellVizConfig] below. The
-// caller may pass a custom instance to override any defaults. Later
-// build steps (see FSD/CELL_VIZ_REDESIGN.md step 11) will plumb a
-// user-facing Settings screen into this config, so the same struct
-// that drives the defaults here is the same struct the user's saved
-// preferences populate at runtime. All fields are hard-capped in
-// [CellVizConfig.sanitized] so a bad preference can't break rendering.
-
-// -----------------------------------------------------------------------------
-// CellVizConfig — every tunable lives here, capped, overridable
-// -----------------------------------------------------------------------------
-
-/**
- * User-facing + developer-facing tuning surface for the cell viz.
- *
- * Every visual constant that a reasonable person might want to change
- * lives on this struct — rotation speed, port size, seam width, future
- * max-node counts, breathing period, gratitude cooldown, etc. Hot-path
- * code reads from a single sanitized instance; the UI-surfaced
- * Settings screen (see build-order step 11) populates the same struct.
- *
- * All numeric fields are clamped to safe ranges by [sanitized]. A user
- * with a broken preferences file can't ship NaN or a negative radius
- * into the renderer — the worst they can do is land at the bounded
- * extreme.
- */
-data class CellVizConfig(
-    // ----- Rotation ---------------------------------------------------------
-    /** Degrees per second for the baseline rotation. 0 disables auto-rotation. */
-    val rotationDegPerSec: Float = 6f,
-
-    // ----- Membrane geometry -----------------------------------------------
-    /**
-     * Membrane radius as a fraction of `min(canvas width, canvas height)`.
-     * 0.26 keeps the full circle inside the chat area even on landscape
-     * layouts where the canvas is shorter than it is wide. Larger values
-     * cause the top/bottom of the cell to clip behind surrounding UI
-     * chrome (warning banner, input bar). Adjustable via Settings in
-     * step 11.
-     */
-    val membraneRadiusFraction: Float = 0.26f,
-
-    /** Stroke width of the bright inner stroke of each bus arc. */
-    val busArcStrokeWidth: Float = 4.5f,
-
-    /** Stroke width of the soft mid-halo stack in dark mode. */
-    val busArcMidHaloWidth: Float = 10f,
-
-    /** Stroke width of the outermost bloom halo in dark mode. */
-    val busArcOuterHaloWidth: Float = 20f,
-
-    // ----- Membrane openings (dynamic apertures) ---------------------------
-    //
-    // The membrane is never a closed ring. At any time, between [minOpenings]
-    // and [maxOpenings] apertures are open somewhere around the cell — each
-    // forms, drifts, and dissolves on its own timer. This renders
-    // Incompleteness Awareness as continuous motion rather than a static
-    // broken pixel that looks like a bug. Openness over stitching.
-
-    /** Minimum number of openings present at any time. */
-    val minOpenings: Int = 3,
-
-    /** Maximum number of openings present at any time. */
-    val maxOpenings: Int = 5,
-
-    /** Seconds an opening takes to grow from 0° to its target width. */
-    val openingGrowSec: Float = 0.8f,
-
-    /** Seconds an opening takes to shrink from its target width back to 0°. */
-    val openingShrinkSec: Float = 0.8f,
-
-    /** Minimum time (seconds) an opening stays at its target width. */
-    val openingStableMinSec: Float = 2.0f,
-
-    /** Maximum time (seconds) an opening stays at its target width. */
-    val openingStableMaxSec: Float = 4.0f,
-
-    /** Minimum angular width (degrees) an opening reaches at full size. */
-    val openingMinWidthDeg: Float = 4f,
-
-    /** Maximum angular width (degrees) an opening reaches at full size. */
-    val openingMaxWidthDeg: Float = 10f,
-
-    /** Maximum drift speed (degrees/sec) an opening walks around the cell. */
-    val openingDriftMaxDegPerSec: Float = 1.0f,
-
-    // ----- Adapter ports ---------------------------------------------------
-    /** Half-extent (in pixels) of an adapter port's shape. */
-    val portRadiusPx: Float = 14f,
-
-    /** Margin (in degrees) between a port and its bus segment boundary. */
-    val portSegmentMarginDeg: Float = 8f,
-
-    /** Alpha multiplier applied to inactive-adapter ports. */
-    val portInactiveAlpha: Float = 0.35f,
-
-    // ----- Future steps (placeholders, already capped) ---------------------
-    // Expose these now so the Settings schema is stable even before the
-    // code that reads them lands. Each one has a sensible default and
-    // a hard-capped sanitized value.
-
-    /** Maximum number of memory-graph motes rendered in cytoplasm (step 5). */
-    val maxMemoryMotes: Int = 200,
-
-    /**
-     * Query period (seconds) for refreshing the memory-graph snapshot
-     * that populates motes. Too-frequent queries harass the server;
-     * too-infrequent queries make the cell feel stale.
-     */
-    val memoryQueryPeriodSec: Float = 15f,
-
-    /**
-     * Hours of recent graph history to load for motes (step 5).
-     * 24 matches the legacy LiveGraphBackground behaviour — last
-     * day's worth of activity.
-     */
-    val memoryLoadWindowHours: Int = 24,
-
-    /** Base radius of a single cytoplasm mote in pixels. */
-    val moteRadiusPx: Float = 2.4f,
-
-    /**
-     * Peak drift amplitude in pixels. Each mote oscillates with its
-     * own phase/frequency within this envelope — small enough to feel
-     * ambient, large enough to register as "alive" (plan §3).
-     */
-    val moteDriftAmpPx: Float = 3f,
-
-    /**
-     * Reference drift period in seconds. Individual motes pick slightly
-     * different frequencies so they don't drift in lockstep.
-     */
-    val moteDriftPeriodSec: Float = 12f,
-
-    /**
-     * Birth animation duration (ms). New motes fade in with a soft
-     * white halo pulse over this window.
-     */
-    val moteBirthMs: Long = 1500L,
-
-    /** Max gratitude motes in flight at once (step 6). */
-    val maxGratitudeMotesInFlight: Int = 1,
-
-    /** Minimum seconds between gratitude-mote emissions (step 6). */
-    val gratitudeCooldownSec: Float = 3f,
-
-    /** Maximum in-flight floating "caught" event bubbles on the UI layer. */
-    val maxCaughtBubbles: Int = 12,
-
-    /** Breathing period in seconds (step 4 — active-presence rhythm). */
-    val breathePeriodSec: Float = 6f,
-
-    /**
-     * Peak scale added on the breathe animation. 0.035 = 3.5% — clearly
-     * perceptible without feeling like the cell is gasping. At 1.8% the
-     * motion was too subtle for a casual observer to register that the
-     * system is active.
-     */
-    val breatheScaleAmp: Float = 0.035f,
-
-    /**
-     * Nucleus song period in seconds. 6 s gives a visible wave
-     * propagating outward every 6 s — unrushed but clearly alive.
-     */
-    val nucleusSongPeriodSec: Float = 6f,
-
-    /**
-     * Nucleus outer radius as a fraction of the membrane radius.
-     * 0.45 makes the pipeline area a meaningful, readable presence at
-     * the cell's centre — large enough that the shells and song waves
-     * are perceptible without competing with the membrane.
-     */
-    val nucleusRadiusFraction: Float = 0.45f,
-
-    /**
-     * Peak opacity of each emitted nucleus song wave. 0.65 with the
-     * three-path bloom stack makes the wave unmistakable against a
-     * dense cytoplasm of motes in dark mode. Drop lower for a more
-     * ambient feel; 1.0 maxes out.
-     */
-    val nucleusSongPeakOpacity: Float = 0.65f,
-
-    /**
-     * Emit a song wave every Nth period. Default 1 (every cycle) at
-     * the default 8 s period gives one wave roughly every 8 s. Raise
-     * to 2 or 3 for a quieter hum.
-     */
-    val nucleusSongEmissionEveryN: Int = 1,
-) {
-    /**
-     * Return a copy with every value forced into a safe range. Call
-     * this once per composition; every tunable is then known to be in
-     * bounds for the rest of the rendering pass.
-     */
-    fun sanitized(): CellVizConfig {
-        // Openings: min ≤ max, min ≥ 0. Keep both in [0, 8] so the
-        // membrane can't become so holey that the cell reads as broken.
-        val sanitizedMin = minOpenings.coerceIn(0, 8)
-        val sanitizedMax = maxOpenings.coerceIn(sanitizedMin, 8)
-        val sanitizedStableMin = openingStableMinSec.coerceIn(0.2f, 20f)
-        val sanitizedStableMax = openingStableMaxSec.coerceIn(sanitizedStableMin, 30f)
-        val sanitizedWidthMin = openingMinWidthDeg.coerceIn(1f, 45f)
-        val sanitizedWidthMax = openingMaxWidthDeg.coerceIn(sanitizedWidthMin, 45f)
-
-        return copy(
-            rotationDegPerSec         = rotationDegPerSec.coerceIn(0f, 45f),
-            membraneRadiusFraction    = membraneRadiusFraction.coerceIn(0.15f, 0.48f),
-            busArcStrokeWidth         = busArcStrokeWidth.coerceIn(1f, 12f),
-            busArcMidHaloWidth        = busArcMidHaloWidth.coerceIn(2f, 24f),
-            busArcOuterHaloWidth      = busArcOuterHaloWidth.coerceIn(4f, 48f),
-            minOpenings               = sanitizedMin,
-            maxOpenings               = sanitizedMax,
-            openingGrowSec            = openingGrowSec.coerceIn(0.1f, 5f),
-            openingShrinkSec          = openingShrinkSec.coerceIn(0.1f, 5f),
-            openingStableMinSec       = sanitizedStableMin,
-            openingStableMaxSec       = sanitizedStableMax,
-            openingMinWidthDeg        = sanitizedWidthMin,
-            openingMaxWidthDeg        = sanitizedWidthMax,
-            openingDriftMaxDegPerSec  = openingDriftMaxDegPerSec.coerceIn(0f, 10f),
-            portRadiusPx              = portRadiusPx.coerceIn(6f, 30f),
-            portSegmentMarginDeg      = portSegmentMarginDeg.coerceIn(0f, 20f),
-            portInactiveAlpha         = portInactiveAlpha.coerceIn(0.1f, 1f),
-            maxMemoryMotes            = maxMemoryMotes.coerceIn(0, 500),
-            memoryQueryPeriodSec      = memoryQueryPeriodSec.coerceIn(3f, 300f),
-            memoryLoadWindowHours     = memoryLoadWindowHours.coerceIn(1, 168),
-            moteRadiusPx              = moteRadiusPx.coerceIn(0.5f, 12f),
-            moteDriftAmpPx            = moteDriftAmpPx.coerceIn(0f, 20f),
-            moteDriftPeriodSec        = moteDriftPeriodSec.coerceIn(2f, 60f),
-            moteBirthMs               = moteBirthMs.coerceIn(0L, 6000L),
-            maxGratitudeMotesInFlight = maxGratitudeMotesInFlight.coerceIn(0, 4),
-            gratitudeCooldownSec      = gratitudeCooldownSec.coerceIn(0.5f, 60f),
-            maxCaughtBubbles          = maxCaughtBubbles.coerceIn(0, 32),
-            breathePeriodSec          = breathePeriodSec.coerceIn(2f, 30f),
-            breatheScaleAmp           = breatheScaleAmp.coerceIn(0f, 0.06f),
-            nucleusSongPeriodSec      = nucleusSongPeriodSec.coerceIn(2f, 30f),
-            nucleusRadiusFraction     = nucleusRadiusFraction.coerceIn(0.10f, 0.60f),
-            nucleusSongPeakOpacity    = nucleusSongPeakOpacity.coerceIn(0f, 1f),
-            nucleusSongEmissionEveryN = nucleusSongEmissionEveryN.coerceIn(1, 10),
-        )
-    }
-
-    companion object {
-        /** Stable reference to the default, already-sanitized config. */
-        val DEFAULT: CellVizConfig = CellVizConfig().sanitized()
-    }
-}
+//   Step 3 (cell skeleton):   bus arcs, adapter ports, membrane openings
+//   Step 4 (nucleus):         static nucleus + breathing
+//   Step 5 (cytoplasm):       golden-angle motes with drift + birth halo
+//   Step 6+ (events):         not yet — rhythmic bus shimmer, gratitude
+//                              motes, deferral ripple, etc. land later.
 
 // -----------------------------------------------------------------------------
 // Bus segments — load-bearing, do not rearrange
 // -----------------------------------------------------------------------------
 
-/** Identifier for one of CIRIS's six message buses. */
-enum class CellBus { COMM, MEMORY, LLM, TOOL, RUNTIME, WISE }
-
 /**
  * Fixed angle + color for a bus arc on the membrane. Angles are in
  * degrees, Compose convention (0° = east, clockwise).
+ *
+ * This data class depends on Compose via [Color], which is why it
+ * lives here rather than in the pure-Kotlin [CellVizModel.kt].
  */
-private data class BusSegment(
+internal data class BusSegment(
     val bus: CellBus,
     val startDeg: Float,
     val endDeg: Float,
@@ -331,9 +65,10 @@ private data class BusSegment(
 
 /**
  * Canonical bus layout — order matters only for iteration; each segment
- * owns a fixed 60° range of the membrane.
+ * owns a fixed 60° range of the membrane. Colours are load-bearing
+ * (see FSD/CELL_VIZ_REDESIGN.md §2) and must not be made user-tunable.
  */
-private val BUS_SEGMENTS: List<BusSegment> = listOf(
+internal val BUS_SEGMENTS: List<BusSegment> = listOf(
     BusSegment(CellBus.TOOL,    startDeg = 0f,   endDeg = 60f,  color = Color(0xFFD98A2D)),
     BusSegment(CellBus.LLM,     startDeg = 60f,  endDeg = 120f, color = Color(0xFF3D86D9)),
     BusSegment(CellBus.MEMORY,  startDeg = 120f, endDeg = 180f, color = Color(0xFF8B5FD6)),
@@ -343,144 +78,35 @@ private val BUS_SEGMENTS: List<BusSegment> = listOf(
 )
 
 // -----------------------------------------------------------------------------
-// Membrane openings — dynamic apertures
+// Nucleus styling
 // -----------------------------------------------------------------------------
-//
-// The membrane is never a closed ring. 3–5 apertures exist at any time;
-// each one grows in, drifts slightly, stabilizes, then shrinks out. When
-// one dies a new one spawns elsewhere. This renders the "openness" and
-// "imperfection" principles as continuous motion, and completely avoids
-// the "looks like a bug" read of a static gap.
-//
-// The openings CUT bus arcs — wherever an opening overlaps a bus, that
-// portion of the arc is not drawn. Adapter ports still render on top
-// because they sit above the arc layer.
+
+/** Warm amber the nucleus emits — fixed, not theme-derived. */
+private val NUCLEUS_AMBER = Color(0xFFE3A64B)
 
 /**
- * One aperture in the cell membrane at a given moment in time.
- *
- * An opening has a deterministic lifecycle: grow → stable → shrink → die.
- * All timing is expressed in wall-clock ms so the renderer can drive its
- * state purely from the current time, without per-frame mutation of the
- * opening itself. Creating a new `MembraneOpening` is the only mutation;
- * everything else is a pure function of age.
+ * Radii of the 7 nucleus shells as fractions of the nucleus outer
+ * radius. Picked to be readable but not crowded; each shell is
+ * slightly thinner than the last as we move outward.
  */
-private data class MembraneOpening(
-    val id: Long,
-    /** Initial position of the opening's center (degrees, 0..360). */
-    val birthCenterDeg: Float,
-    /** Target angular width at full growth (degrees). */
-    val targetWidthDeg: Float,
-    /** How fast the opening walks around the cell (deg/sec; can be negative). */
-    val driftDegPerSec: Float,
-    /** Wall-clock millis when this opening was born. */
-    val bornAtMs: Long,
-    /** Grow-in duration from 0 → targetWidthDeg (ms). */
-    val growMs: Long,
-    /** Time at full width (ms). */
-    val stableMs: Long,
-    /** Shrink-out duration back to 0 (ms). */
-    val shrinkMs: Long,
-) {
-    /** Total lifetime in ms (grow + stable + shrink). */
-    val lifetimeMs: Long get() = growMs + stableMs + shrinkMs
-
-    fun isDead(nowMs: Long): Boolean = nowMs >= bornAtMs + lifetimeMs
-
-    /** Current center angle in degrees, wrapped to 0..360. */
-    fun currentCenterDeg(nowMs: Long): Float {
-        val ageSec = (nowMs - bornAtMs) / 1000f
-        val raw = birthCenterDeg + driftDegPerSec * ageSec
-        return ((raw % 360f) + 360f) % 360f
-    }
-
-    /** Current angular width in degrees — 0 before birth, 0 after death. */
-    fun currentWidthDeg(nowMs: Long): Float {
-        val age = nowMs - bornAtMs
-        return when {
-            age < 0L -> 0f
-            age < growMs -> targetWidthDeg * smoothstep(age.toFloat() / growMs)
-            age < growMs + stableMs -> targetWidthDeg
-            age < lifetimeMs -> targetWidthDeg * smoothstep(
-                (lifetimeMs - age).toFloat() / shrinkMs
-            )
-            else -> 0f
-        }
-    }
-}
-
-/** Classic smoothstep easing — C¹ continuous, no library dependency. */
-private fun smoothstep(t: Float): Float {
-    val x = t.coerceIn(0f, 1f)
-    return x * x * (3f - 2f * x)
-}
-
-// -----------------------------------------------------------------------------
-// Cytoplasm motes — the memory graph, visible in the cell
-// -----------------------------------------------------------------------------
-//
-// Every node in the last [cfg.memoryLoadWindowHours] hours of the memory
-// graph renders as a small luminous mote drifting in the cytoplasm —
-// the region between the nucleus and the membrane. Positions are
-// deterministic functions of (stable index, time); no per-frame
-// mutation of mote fields.
-//
-// Positioning uses the golden-angle scatter (φ ≈ 137.508°) so motes
-// distribute evenly without visible banding or clustering. Radial
-// distance is sqrt-based so each equal-area ring contains roughly the
-// same number of motes — uniform 2D density, not concentrated near the
-// nucleus.
-
-/**
- * One node from the memory graph, laid out as a drifting mote.
- *
- * The `stableIndex` is what determines this mote's angular + radial
- * placement via the golden-angle formula. Preserving the same index
- * across refreshes means a node doesn't "teleport" around the cell
- * when neighbouring nodes appear or disappear — it breathes in place.
- */
-private data class CytoplasmMote(
-    val id: String,
-    val scope: ai.ciris.api.models.GraphScope,
-    val stableIndex: Int,
-    val birthTimeMs: Long,
+private val NUCLEUS_SHELL_FRACTIONS = floatArrayOf(
+    0.25f, 0.35f, 0.45f, 0.55f, 0.65f, 0.78f, 0.92f,
 )
 
-/**
- * Map an adapter's type string onto the bus that owns it. Step-3 only
- * needs the 1-to-1 type→bus mapping; per-adapter positioning within a
- * bus segment comes from spread.
- */
-private fun adapterBus(type: String): CellBus = when (type.lowercase()) {
-    "api", "discord", "cli" -> CellBus.COMM
-    "weather", "navigation", "home_assistant", "ha",
-    "wallet", "reddit", "mcp", "apple_notes", "apple_reminders",
-    "bear_notes", "bird", "blogwatcher" -> CellBus.TOOL
-    "cirisverify" -> CellBus.WISE
-    // LLM, MEMORY, RUNTIME don't typically come from "adapter" entries
-    // in the current data model. Default to TOOL as a visible placeholder.
-    else -> CellBus.TOOL
-}
+/** Shell opacities, matched index-by-index to [NUCLEUS_SHELL_FRACTIONS]. */
+private val NUCLEUS_SHELL_OPACITIES = floatArrayOf(
+    0.40f, 0.42f, 0.42f, 0.38f, 0.32f, 0.24f, 0.16f,
+)
 
-/** Which rendered shape a bus uses for its adapter ports. */
-private enum class PortShape { DIAMOND, HEX }
-
-private fun portShapeFor(bus: CellBus): PortShape = when (bus) {
-    // Memory and graph-shaped buses use hex; flow buses use diamond.
-    // Keeps the vocabulary consistent with the mockup.
-    CellBus.MEMORY, CellBus.WISE -> PortShape.HEX
-    else -> PortShape.DIAMOND
-}
-
-// -----------------------------------------------------------------------------
+// =============================================================================
 // Public composable
-// -----------------------------------------------------------------------------
+// =============================================================================
 
 /**
- * Cell visualization. Step 3 of the redesign — renders the medium, the
- * membrane (bus arcs with dynamic openings), and adapter ports anchored
- * on their owning bus. Drop-in replacement for [LiveGraphBackground] at
- * call sites that passed the [probeCellVizCapability] gate.
+ * Cell visualization. Renders the medium, membrane (bus arcs with
+ * dynamic openings), cytoplasm motes, adapter ports, and nucleus.
+ * Drop-in replacement for [LiveGraphBackground] at call sites that
+ * passed the `probeCellVizCapability` gate.
  *
  * @param modifier Layout modifier — fill the chat area; the cell will
  *   center itself within the available space.
@@ -492,10 +118,17 @@ private fun portShapeFor(bus: CellBus): PortShape = when (bus) {
  *   owning bus segment, not on arbitrary altitudes.
  * @param externalRotation Degrees added to the baseline rotation
  *   (swipe-to-spin).
- * @param config Every tunable the viz exposes. See [CellVizConfig]; the
- *   config is sanitized once per composition so draw code can assume
- *   every field is in a safe range. User-facing Settings screen (build
- *   step 11) will inject overrides here.
+ * @param config Every tunable the viz exposes. See [CellVizConfig];
+ *   the config is sanitized once per composition so draw code can
+ *   assume every field is in a safe range.
+ * @param apiClient Optional — when present, cytoplasm motes populate
+ *   from the live memory graph. Null means "render the cell empty of
+ *   motes", which is a legitimate pre-login state.
+ * @param colorTheme Used only to color memory motes via
+ *   [getScopeColor]. Bus / adapter / nucleus colours are architectural.
+ * @param eventTrigger Incremented by the caller when something happens
+ *   that might have changed the memory graph. Step 5 fetches once at
+ *   mount; step 6+ makes this reactive.
  */
 @Composable
 fun CellVisualization(
@@ -504,27 +137,17 @@ fun CellVisualization(
     adapterOrbits: List<AdapterOrbit> = emptyList(),
     externalRotation: Float = 0f,
     config: CellVizConfig = CellVizConfig.DEFAULT,
-    // Optional — when present, cytoplasm motes populate from the live
-    // memory graph. Null means "render the cell empty of motes", which
-    // is a legitimate state (e.g. during startup) and should not crash.
     apiClient: ai.ciris.mobile.shared.api.CIRISApiClient? = null,
     colorTheme: ai.ciris.mobile.shared.ui.theme.ColorTheme =
         ai.ciris.mobile.shared.ui.theme.ColorTheme.DEFAULT,
-    // Incremented by the caller when something happened that might have
-    // changed the memory graph (SSE events, explicit refresh). Wired in
-    // step 6; step 5 just fetches once at mount.
     eventTrigger: Int = 0,
 ) {
-    // Take the config once per composition (effectively per param change)
-    // so every draw reads the same sanitized values.
+    // Sanitize once per composition so draw code reads in-range values.
     val cfg = remember(config) { config.sanitized() }
 
-    // --- Rotation driver (matches step-2 refactor exactly) ---------------
-    //
-    // One mutable float; one LaunchedEffect driving it with withFrameNanos.
-    // Delta-time accumulation gives the canonical revolution without
-    // relying on a tween spec that can't be paused.
-
+    // Rotation driver: withFrameNanos accumulates delta-time. Single
+    // source of truth for current angle; the membrane, ports, and any
+    // future rotating element all derive from it.
     var autoRotationDeg by remember { mutableStateOf(0f) }
     LaunchedEffect(cfg.rotationDegPerSec) {
         var lastFrameNs = 0L
@@ -532,7 +155,8 @@ fun CellVisualization(
             withFrameNanos { frameTimeNs ->
                 if (lastFrameNs != 0L) {
                     val dSec = (frameTimeNs - lastFrameNs) / 1_000_000_000f
-                    autoRotationDeg = (autoRotationDeg + cfg.rotationDegPerSec * dSec) % 360f
+                    autoRotationDeg =
+                        (autoRotationDeg + cfg.rotationDegPerSec * dSec) % 360f
                 }
                 lastFrameNs = frameTimeNs
             }
@@ -540,19 +164,19 @@ fun CellVisualization(
     }
     val rotationDeg = (autoRotationDeg + externalRotation) % 360f
 
-    // --- Membrane openings --------------------------------------------------
-    //
-    // We keep a list of live openings in Compose state. A frame-timer
-    // effect (a) expires dead openings and (b) spawns new ones to keep the
-    // count at [cfg.minOpenings]. Current center/width are PURE functions of
-    // wall-clock time — per-frame draw just reads them, no mutation.
-
+    // Membrane openings: frame-timer effect expires dead openings and
+    // spawns new ones to stay within [cfg.minOpenings, cfg.maxOpenings].
+    // Current center/width are PURE functions of wall-clock time — per-frame
+    // draw just reads them, no mutation of opening fields.
     val openings = remember { mutableStateOf(emptyList<MembraneOpening>()) }
     val nowMs = remember { mutableStateOf(0L) }
 
-    LaunchedEffect(cfg.minOpenings, cfg.maxOpenings, cfg.openingStableMinSec,
-        cfg.openingStableMaxSec, cfg.openingMinWidthDeg, cfg.openingMaxWidthDeg,
-        cfg.openingDriftMaxDegPerSec, cfg.openingGrowSec, cfg.openingShrinkSec) {
+    LaunchedEffect(
+        cfg.minOpenings, cfg.maxOpenings,
+        cfg.openingStableMinSec, cfg.openingStableMaxSec,
+        cfg.openingMinWidthDeg, cfg.openingMaxWidthDeg,
+        cfg.openingDriftMaxDegPerSec, cfg.openingGrowSec, cfg.openingShrinkSec,
+    ) {
         var lastFrameNs = 0L
         val rng = kotlin.random.Random.Default
         while (isActive) {
@@ -562,12 +186,10 @@ fun CellVisualization(
                     nowMs.value = now
                     val alive = openings.value.filterNot { it.isDead(now) }
                     val delta = mutableListOf<MembraneOpening>()
-                    // Spawn until we hit the minimum; above that, let
-                    // natural death thin out and only spawn probabilistically.
                     if (alive.size < cfg.minOpenings) {
                         delta += spawnOpening(now, rng, cfg)
                     } else if (alive.size < cfg.maxOpenings && rng.nextFloat() < 0.002f) {
-                        // Very small per-frame probability → ~0.12/sec at 60fps
+                        // ~0.12/sec at 60fps; gentle top-up above the minimum.
                         delta += spawnOpening(now, rng, cfg)
                     }
                     if (delta.isNotEmpty() || alive.size != openings.value.size) {
@@ -579,25 +201,20 @@ fun CellVisualization(
         }
     }
 
-    // --- Group adapters by bus so ports can spread within each segment ---
+    // Group adapters by bus so ports can spread within each segment.
     val adaptersByBus: Map<CellBus, List<AdapterOrbit>> = remember(adapterOrbits) {
         adapterOrbits.groupBy { adapterBus(it.type) }
     }
 
-    // --- Cytoplasm motes (memory graph) ---------------------------------
-    //
-    // Fetch once at mount + whenever eventTrigger changes. Each mote
-    // keeps a stableIndex across refreshes so existing nodes don't
-    // teleport when neighbours appear or disappear — they just fade in
-    // (birth halo) or fade out (removed). Total count bounded by
-    // cfg.maxMemoryMotes; server-side `limit` caps the fetch matching.
+    // Cytoplasm motes (memory graph). Fetch on mount + eventTrigger change.
+    // stableIndex preserved across refreshes so motes don't teleport.
     val motes = remember { mutableStateOf(emptyList<CytoplasmMote>()) }
-    // Preserved indices for stable positioning across refreshes.
     val moteIndexById = remember { mutableMapOf<String, Int>() }
     val nextMoteIndex = remember { mutableStateOf(0) }
 
-    LaunchedEffect(apiClient, eventTrigger, cfg.maxMemoryMotes,
-        cfg.memoryLoadWindowHours) {
+    LaunchedEffect(
+        apiClient, eventTrigger, cfg.maxMemoryMotes, cfg.memoryLoadWindowHours,
+    ) {
         val client = apiClient ?: return@LaunchedEffect
         try {
             val graph = client.getGraphData(
@@ -608,14 +225,14 @@ fun CellVisualization(
                 includeMetrics = false,
             )
             val nowMsLocal = kotlinx.datetime.Clock.System.now().toEpochMilliseconds()
-            val known = moteIndexById.keys.toSet()
+            val wasPopulated = moteIndexById.isNotEmpty()
             val newMotes = graph.nodes.map { node ->
                 val existingIdx = moteIndexById[node.id]
                 val idx = existingIdx ?: nextMoteIndex.value.also {
                     moteIndexById[node.id] = it
                     nextMoteIndex.value = it + 1
                 }
-                val isNew = existingIdx == null && known.isNotEmpty()
+                val isNew = existingIdx == null && wasPopulated
                 CytoplasmMote(
                     id = node.id,
                     scope = node.scope,
@@ -638,33 +255,28 @@ fun CellVisualization(
         val membraneRadius = minOf(size.width, size.height) * cfg.membraneRadiusFraction
         val nucleusRadius = membraneRadius * cfg.nucleusRadiusFraction
 
-        // Wall-clock-driven scalars.
-        //
-        // breathePhase is a continuous phase in radians that advances with
-        // time; the scale multiplier is a small sin-wave around 1.0. Aura
-        // opacity pulses in phase so "the cell body becomes slightly
-        // brighter as it expands" — the two cues reinforce each other.
+        // Breathing: whole-cell scale + aura opacity pulse, both driven
+        // by the same sin-wave phase so the two cues reinforce.
         val nowSec = nowMs.value / 1000f
         val breathePhase = (nowSec / cfg.breathePeriodSec) * 2f * PI.toFloat()
         val breatheScale = 1f + cfg.breatheScaleAmp * sin(breathePhase)
         val breatheAuraAlpha = 0.85f + (cfg.breatheScaleAmp / 0.010f) * 0.15f *
-            (0.5f * (1f + sin(breathePhase)))  // 0.85..1.00 when amp=0.010
+            (0.5f * (1f + sin(breathePhase)))
 
         drawMedium(isDarkMode, centerX, centerY)
 
-        // Everything that IS the cell (not the medium around it) breathes
-        // together. A single scale transform applies uniformly.
-        scale(scaleX = breatheScale, scaleY = breatheScale,
-            pivot = Offset(centerX, centerY)) {
+        // Everything inside the cell scales uniformly; medium does not.
+        scale(
+            scaleX = breatheScale, scaleY = breatheScale,
+            pivot = Offset(centerX, centerY),
+        ) {
             drawCellBodyAura(
                 isDarkMode = isDarkMode,
                 cx = centerX, cy = centerY,
                 radius = membraneRadius * 1.05f,
                 opacityMultiplier = breatheAuraAlpha,
             )
-            // Collect current opening ranges once per frame; pass through to
-            // the membrane draw which subtracts them from the bus arcs.
-            val openingRanges = openings.value
+            val openingAngleRanges = openings.value
                 .flatMap { openingRanges(it, nowMs.value) }
             drawMembrane(
                 rotationDeg = rotationDeg,
@@ -672,7 +284,7 @@ fun CellVisualization(
                 radius = membraneRadius,
                 isDarkMode = isDarkMode,
                 cfg = cfg,
-                openingRanges = openingRanges,
+                openingRanges = openingAngleRanges,
             )
             drawCytoplasmMotes(
                 motes = motes.value,
@@ -697,62 +309,13 @@ fun CellVisualization(
                 outerRadius = nucleusRadius,
                 isDarkMode = isDarkMode,
             )
-            drawNucleusSong(
-                cx = centerX, cy = centerY,
-                nucleusOuterRadius = nucleusRadius,
-                membraneRadius = membraneRadius,
-                isDarkMode = isDarkMode,
-                cfg = cfg,
-                nowSec = nowSec,
-            )
+
+            // The nucleus "song" (slow pulse wave from the core through
+            // the cytoplasm) was removed after repeated tuning attempts
+            // failed to make it reliably read as *moving* against the
+            // dense mote cloud. Rhythmic signalling is now carried by
+            // breathing + the per-event pulses landing in step 6.
         }
-    }
-}
-
-/**
- * Build a new opening with randomized width, drift, and stable duration
- * picked from the configured ranges. Spawn position is uniform around
- * the cell so openings don't cluster.
- */
-private fun spawnOpening(
-    nowMs: Long,
-    rng: kotlin.random.Random,
-    cfg: CellVizConfig,
-): MembraneOpening {
-    val widthDeg = cfg.openingMinWidthDeg +
-        rng.nextFloat() * (cfg.openingMaxWidthDeg - cfg.openingMinWidthDeg)
-    val stableSec = cfg.openingStableMinSec +
-        rng.nextFloat() * (cfg.openingStableMaxSec - cfg.openingStableMinSec)
-    val drift = (rng.nextFloat() - 0.5f) * 2f * cfg.openingDriftMaxDegPerSec
-    return MembraneOpening(
-        id = rng.nextLong(),
-        birthCenterDeg = rng.nextFloat() * 360f,
-        targetWidthDeg = widthDeg,
-        driftDegPerSec = drift,
-        bornAtMs = nowMs,
-        growMs = (cfg.openingGrowSec * 1000).toLong(),
-        stableMs = (stableSec * 1000).toLong(),
-        shrinkMs = (cfg.openingShrinkSec * 1000).toLong(),
-    )
-}
-
-/**
- * Return the degree ranges currently occluded by an opening. A seam
- * straddling the 0°/360° boundary returns two ranges; all others return
- * one. An opening with zero current width (freshly spawned or just
- * dying) returns an empty list.
- */
-private fun openingRanges(op: MembraneOpening, nowMs: Long): List<ClosedFloatingPointRange<Float>> {
-    val w = op.currentWidthDeg(nowMs)
-    if (w <= 0f) return emptyList()
-    val c = op.currentCenterDeg(nowMs)
-    val half = w / 2f
-    val rawStart = c - half
-    val rawEnd = c + half
-    return when {
-        rawStart < 0f -> listOf(0f..rawEnd, (rawStart + 360f)..360f)
-        rawEnd > 360f -> listOf(rawStart..360f, 0f..(rawEnd - 360f))
-        else -> listOf(rawStart..rawEnd)
     }
 }
 
@@ -784,11 +347,9 @@ private fun DrawScope.drawMedium(isDarkMode: Boolean, cx: Float, cy: Float) {
 
 /**
  * A faint warm glow centered on the cell, so you register "there's a
- * body here" without naming it. Intentionally almost invisible.
- *
- * [opacityMultiplier] is the breathe-driven opacity modulation: the aura
- * brightens and dims in step with the cell's scale pulse so the two cues
- * reinforce each other rather than fighting.
+ * body here" without naming it. [opacityMultiplier] is the breathe-
+ * driven opacity modulation — the aura brightens and dims in step
+ * with the cell's scale pulse so the two cues reinforce.
  */
 private fun DrawScope.drawCellBodyAura(
     isDarkMode: Boolean,
@@ -815,187 +376,11 @@ private fun DrawScope.drawCellBodyAura(
     )
 }
 
-// =============================================================================
-// Nucleus + song — the cell's small, warm center
-// =============================================================================
-//
-// The pipeline lives here, not around the whole membrane. It's small
-// (cfg.nucleusRadiusFraction × membraneRadius, default 30%) and sits
-// dead-center. Seven thin concentric shells represent the H3ERE stages
-// (THINK..ACT); the shells do not individually animate in step 4, they
-// are the static anatomy. A slow "song" wave emits from the core every
-// Nth cycle — not a heartbeat, a hum.
-
 /**
- * Radii of the 7 nucleus shells as fractions of the nucleus outer
- * radius. Picked to be readable but not crowded; each shell is slightly
- * thinner than the last as we move outward.
- */
-private val NUCLEUS_SHELL_FRACTIONS = floatArrayOf(
-    0.25f, 0.35f, 0.45f, 0.55f, 0.65f, 0.78f, 0.92f,
-)
-
-/** Shell opacities, matched index-by-index to [NUCLEUS_SHELL_FRACTIONS]. */
-private val NUCLEUS_SHELL_OPACITIES = floatArrayOf(
-    0.40f, 0.42f, 0.42f, 0.38f, 0.32f, 0.24f, 0.16f,
-)
-
-/** Warm amber the nucleus emits — fixed, not theme-derived. */
-private val NUCLEUS_AMBER = Color(0xFFE3A64B)
-
-/**
- * Draw the nucleus — a warm radial-gradient fill, 7 concentric shells,
- * and a bright emissive core at the centre. Zero per-frame state; the
- * nucleus is just static anatomy until a pipeline event lights it up
- * (that wiring comes in step 6).
- */
-private fun DrawScope.drawNucleus(
-    cx: Float,
-    cy: Float,
-    outerRadius: Float,
-    isDarkMode: Boolean,
-) {
-    if (outerRadius <= 1f) return
-    val center = Offset(cx, cy)
-
-    // Warm fill — a soft amber wash. We deliberately do NOT run pure
-    // white-amber at full opacity in the centre; the earlier version
-    // read as eye-searing against the indigo-black medium. Amber-only
-    // reads warm without hurting to look at.
-    val fillInner = NUCLEUS_AMBER.copy(alpha = if (isDarkMode) 0.38f else 0.25f)
-    val fillMid   = NUCLEUS_AMBER.copy(alpha = if (isDarkMode) 0.22f else 0.15f)
-    val fillOuter = NUCLEUS_AMBER.copy(alpha = 0f)
-    drawCircle(
-        brush = Brush.radialGradient(
-            colors = listOf(fillInner, fillMid, fillOuter),
-            center = center,
-            radius = outerRadius,
-        ),
-        radius = outerRadius,
-        center = center,
-    )
-
-    // Seven shells, thin strokes, slightly warmer amber than the fill so
-    // they read against the gradient.
-    val shellColor = NUCLEUS_AMBER
-    NUCLEUS_SHELL_FRACTIONS.forEachIndexed { i, frac ->
-        val r = outerRadius * frac
-        drawCircle(
-            color = shellColor.copy(alpha = NUCLEUS_SHELL_OPACITIES[i]),
-            radius = r,
-            center = center,
-            style = Stroke(width = 0.9f),
-        )
-    }
-
-    // Soft inner core. Amber (not white), no hard-edge core dot. A
-    // small warm centre of light that doesn't punch out of the scene.
-    val coreRadius = outerRadius * 0.10f
-    drawCircle(
-        color = NUCLEUS_AMBER.copy(alpha = if (isDarkMode) 0.35f else 0.25f),
-        radius = coreRadius * 2.2f,
-        center = center,
-    )
-    drawCircle(
-        color = NUCLEUS_AMBER.copy(alpha = if (isDarkMode) 0.70f else 0.55f),
-        radius = coreRadius,
-        center = center,
-    )
-}
-
-/**
- * The nucleus "song" — a slow concentric wave emitted from the core
- * every [CellVizConfig.nucleusSongEmissionEveryN] periods. The wave
- * propagates from inside the nucleus outward through the cytoplasm,
- * nearly reaching the membrane before dissolving.
- *
- * Design tuning:
- *  - The wave travels WELL PAST the nucleus boundary (maxR = 0.85 ×
- *    membraneRadius) so it's visible as an expanding ring against the
- *    dark cytoplasm, not trapped inside the nucleus fill's amber glow.
- *  - Stroke is thick (3.5 px in dark mode) so the ring reads clearly
- *    at the low peak opacity.
- *  - Envelope is half-sine so the wave swells in and fades out over
- *    its cycle rather than hard-starting and hard-ending.
- */
-private fun DrawScope.drawNucleusSong(
-    cx: Float,
-    cy: Float,
-    nucleusOuterRadius: Float,
-    membraneRadius: Float,
-    isDarkMode: Boolean,
-    cfg: CellVizConfig,
-    nowSec: Float,
-) {
-    if (cfg.nucleusSongPeakOpacity <= 0f) return
-    val period = cfg.nucleusSongPeriodSec
-    if (period <= 0f) return
-
-    // Which cycle are we in, and are we on an emission cycle?
-    val cycleCount = (nowSec / period).toInt()
-    if (cycleCount % cfg.nucleusSongEmissionEveryN != 0) return
-
-    val phase = ((nowSec % period) / period).coerceIn(0f, 1f)  // 0..1 within the cycle
-    // Half-sine envelope: ramps up, peaks at mid-cycle, ramps down.
-    val envelope = sin(phase * PI.toFloat())  // 0 → 1 → 0
-    val alphaScale = envelope.coerceIn(0f, 1f)
-    if (alphaScale < 0.02f) return
-
-    // Wave grows from the nucleus's inner region out toward the
-    // membrane — it travels through the cytoplasm, not just within the
-    // nucleus. This is what makes the pulse legible: the ring moves
-    // against the dark medium rather than being lost in the nucleus fill.
-    val minR = nucleusOuterRadius * 0.20f
-    val maxR = membraneRadius * 0.85f
-    val r = minR + (maxR - minR) * phase
-
-    // Three-path bloom stack so the wave is unambiguous against a busy
-    // cytoplasm of motes. Two soft halos behind a bright core stroke.
-    // Against a thin stroke alone, motes visually drown out the ring —
-    // bloom widens the contrast zone around the core without blowing
-    // out brightness.
-    val center = Offset(cx, cy)
-    val alpha = (cfg.nucleusSongPeakOpacity * alphaScale).coerceIn(0f, 1f)
-    if (isDarkMode) {
-        drawCircle(
-            color = NUCLEUS_AMBER.copy(alpha = alpha * 0.22f),
-            radius = r, center = center,
-            style = Stroke(width = 22f),
-        )
-        drawCircle(
-            color = NUCLEUS_AMBER.copy(alpha = alpha * 0.45f),
-            radius = r, center = center,
-            style = Stroke(width = 11f),
-        )
-        drawCircle(
-            color = NUCLEUS_AMBER.copy(alpha = alpha),
-            radius = r, center = center,
-            style = Stroke(width = 4.5f),
-        )
-    } else {
-        drawCircle(
-            color = NUCLEUS_AMBER.copy(alpha = alpha),
-            radius = r, center = center,
-            style = Stroke(width = 2.5f),
-        )
-    }
-}
-
-/**
- * The membrane — six bus arcs around the cell, minus wherever a
- * membrane opening is currently punched through them.
- *
- * Rendering strategy (keep it simple, keep it obvious):
- *  1. For each bus segment, subtract all current opening ranges to
- *     produce a list of un-occluded sub-arcs.
- *  2. Draw each sub-arc as either a three-path bloom stack (dark
- *     mode) or a single stroke (light mode).
- *
- * Rotation is applied uniformly to every arc; each segment's own
- * start angle is offset by the current rotation so the cell appears
- * to spin as a solid body. Openings are in absolute screen-degrees,
- * so they stay put while the cell rotates through them — a drifting
- * hole that the membrane slides past, exactly as we want.
+ * The membrane — six bus arcs, minus wherever a membrane opening is
+ * currently punched through them. In dark mode each arc renders as a
+ * three-path bloom stack (outer halo, mid halo, bright stroke); light
+ * mode is a single stroke to keep the diagrammatic feel.
  */
 private fun DrawScope.drawMembrane(
     rotationDeg: Float,
@@ -1007,18 +392,13 @@ private fun DrawScope.drawMembrane(
     openingRanges: List<ClosedFloatingPointRange<Float>>,
 ) {
     BUS_SEGMENTS.forEach { seg ->
-        // Segment spans are in the cell's body frame; add rotation to
-        // convert into the same absolute-screen-degrees space that
-        // opening ranges live in, then subtract any overlaps.
         val segStart = (seg.startDeg + rotationDeg) % 360f
         val segEnd   = (seg.endDeg   + rotationDeg) % 360f
-
         val subArcs = subtractRangesFromArc(segStart, segEnd, openingRanges)
 
         subArcs.forEach { (subStart, subSweep) ->
-            if (subSweep <= 0.5f) return@forEach  // skip hair-thin slivers
+            if (subSweep <= 0.5f) return@forEach
             if (isDarkMode) {
-                // Three-path bloom stack — cheapest form of glow that still reads.
                 drawBusArc(cx, cy, radius, subStart, subSweep,
                     seg.color.copy(alpha = 0.22f), cfg.busArcOuterHaloWidth)
                 drawBusArc(cx, cy, radius, subStart, subSweep,
@@ -1033,51 +413,7 @@ private fun DrawScope.drawMembrane(
     }
 }
 
-/**
- * Subtract a list of opening ranges (absolute degrees, possibly wrapping
- * past 0/360) from a single bus-arc range.
- *
- * Returned sub-arcs are expressed as `(startDeg, sweepDeg)` pairs ready
- * to hand to [drawBusArc]. The arc itself may wrap past 0/360 — in that
- * case we decompose the arc into two non-wrapping halves first, run the
- * subtraction on each, and concatenate. This keeps the inner math simple
- * (closed intervals on [0,360]) at the cost of at most one extra split.
- */
-private fun subtractRangesFromArc(
-    arcStart: Float,
-    arcEnd: Float,
-    openings: List<ClosedFloatingPointRange<Float>>,
-): List<Pair<Float, Float>> {
-    // Decompose the arc into up to two non-wrapping pieces.
-    val arcPieces: List<Pair<Float, Float>> =
-        if (arcEnd >= arcStart) listOf(arcStart to arcEnd)
-        else listOf(arcStart to 360f, 0f to arcEnd)
-
-    val result = mutableListOf<Pair<Float, Float>>()
-    for ((pStart, pEnd) in arcPieces) {
-        val overlaps = openings
-            .mapNotNull { r ->
-                val s = r.start.coerceIn(pStart, pEnd)
-                val e = r.endInclusive.coerceIn(pStart, pEnd)
-                if (e > s) s to e else null
-            }
-            .sortedBy { it.first }
-
-        var cursor = pStart
-        for ((oStart, oEnd) in overlaps) {
-            if (oStart > cursor) result += cursor to oStart
-            cursor = maxOf(cursor, oEnd)
-        }
-        if (cursor < pEnd) result += cursor to pEnd
-    }
-    // Convert (start, end) to (start, sweep)
-    return result.map { (s, e) -> s to (e - s) }
-}
-
-/**
- * Primitive: one arc stroke at a given angle/width/color.
- * Uses the canvas's drawArc on a square bounding box around the cell.
- */
+/** One arc stroke at a given angle/width/color. */
 private fun DrawScope.drawBusArc(
     cx: Float,
     cy: Float,
@@ -1100,13 +436,9 @@ private fun DrawScope.drawBusArc(
 
 /**
  * Adapter ports — diamonds and hexagons anchored on their owning bus
- * arc. Within a bus segment, ports spread evenly with a fixed margin
- * from each segment boundary, so a single-adapter bus sits centered
- * and a three-adapter bus spreads across the arc without touching its
- * neighbors.
- *
- * In dark mode each port gets a halo + core bloom (two extra circles).
- * Inactive adapters render at 35% alpha.
+ * arc. Within a bus segment, ports spread evenly with a margin from
+ * each segment boundary so a single-adapter bus sits centered and a
+ * multi-adapter bus spreads without touching its neighbours.
  */
 private fun DrawScope.drawAdapterPorts(
     adaptersByBus: Map<CellBus, List<AdapterOrbit>>,
@@ -1122,10 +454,15 @@ private fun DrawScope.drawAdapterPorts(
         if (adapters.isEmpty()) return@forEach
 
         adapters.forEachIndexed { index, adapter ->
-            val angleDeg = spreadAngle(seg, index, adapters.size, cfg.portSegmentMarginDeg)
+            val angleDeg = spreadAngle(
+                segmentStartDeg = seg.startDeg,
+                segmentEndDeg = seg.endDeg,
+                index = index,
+                total = adapters.size,
+                marginDeg = cfg.portSegmentMarginDeg,
+            )
             val effectiveDeg = (angleDeg + rotationDeg) % 360f
             val pos = polar(centerX, centerY, membraneRadius, effectiveDeg)
-
             val alpha = if (adapter.isActive) 1f else cfg.portInactiveAlpha
             drawPort(
                 center = pos,
@@ -1140,29 +477,7 @@ private fun DrawScope.drawAdapterPorts(
     }
 }
 
-/**
- * Spread N ports evenly across a segment's arc with a margin from each
- * boundary. For N=1 the port sits at the segment midpoint; for N>=2 the
- * first port is at startDeg+margin and the last at endDeg-margin.
- *
- * Openings in the membrane are dynamic and may briefly cross a port;
- * that's fine — the port renders on top of the arc layer, so a drifting
- * aperture just gives the impression that the port is being passed by.
- */
-private fun spreadAngle(seg: BusSegment, index: Int, total: Int, marginDeg: Float): Float {
-    val usable = seg.sweepDeg - 2f * marginDeg
-    return when {
-        total <= 1 -> seg.midDeg
-        else       -> seg.startDeg + marginDeg + (index.toFloat() / (total - 1)) * usable
-    }
-}
-
-/**
- * Draw one adapter port — core shape + optional halo in dark mode.
- * The port is oriented so the shape's "flat side" faces the center of
- * the cell, which reads as a docked connector rather than a floating
- * badge.
- */
+/** Draw one adapter port — shape + optional halo in dark mode. */
 private fun DrawScope.drawPort(
     center: Offset,
     radius: Float,
@@ -1178,15 +493,13 @@ private fun DrawScope.drawPort(
     }
 
     if (isDarkMode) {
-        // Bloom halo behind the port — two softer circles stacked.
-        drawCircle(color = color.copy(alpha = 0.18f * alpha), radius = radius * 2.0f, center = center)
-        drawCircle(color = color.copy(alpha = 0.32f * alpha), radius = radius * 1.25f, center = center)
+        drawCircle(color = color.copy(alpha = 0.18f * alpha),
+            radius = radius * 2.0f, center = center)
+        drawCircle(color = color.copy(alpha = 0.32f * alpha),
+            radius = radius * 1.25f, center = center)
     }
 
-    drawPath(
-        path = shapePath,
-        color = color.copy(alpha = alpha),
-    )
+    drawPath(path = shapePath, color = color.copy(alpha = alpha))
     drawPath(
         path = shapePath,
         color = (if (isDarkMode) Color.White else Color.Black).copy(alpha = 0.35f * alpha),
@@ -1195,23 +508,69 @@ private fun DrawScope.drawPort(
 }
 
 /**
+ * Nucleus — warm amber fill + 7 concentric shells + soft core.
+ *
+ * Colours are deliberately amber-only (no pure white) so the centre
+ * doesn't read as eye-searing against the indigo-black dark medium.
+ * The shells are the H3ERE pipeline stages as anatomy; individual
+ * shell activation on events lands in step 6.
+ */
+private fun DrawScope.drawNucleus(
+    cx: Float,
+    cy: Float,
+    outerRadius: Float,
+    isDarkMode: Boolean,
+) {
+    if (outerRadius <= 1f) return
+    val center = Offset(cx, cy)
+
+    val fillInner = NUCLEUS_AMBER.copy(alpha = if (isDarkMode) 0.38f else 0.25f)
+    val fillMid   = NUCLEUS_AMBER.copy(alpha = if (isDarkMode) 0.22f else 0.15f)
+    val fillOuter = NUCLEUS_AMBER.copy(alpha = 0f)
+    drawCircle(
+        brush = Brush.radialGradient(
+            colors = listOf(fillInner, fillMid, fillOuter),
+            center = center,
+            radius = outerRadius,
+        ),
+        radius = outerRadius,
+        center = center,
+    )
+
+    NUCLEUS_SHELL_FRACTIONS.forEachIndexed { i, frac ->
+        drawCircle(
+            color = NUCLEUS_AMBER.copy(alpha = NUCLEUS_SHELL_OPACITIES[i]),
+            radius = outerRadius * frac,
+            center = center,
+            style = Stroke(width = 0.9f),
+        )
+    }
+
+    val coreRadius = outerRadius * 0.10f
+    drawCircle(
+        color = NUCLEUS_AMBER.copy(alpha = if (isDarkMode) 0.35f else 0.25f),
+        radius = coreRadius * 2.2f,
+        center = center,
+    )
+    drawCircle(
+        color = NUCLEUS_AMBER.copy(alpha = if (isDarkMode) 0.70f else 0.55f),
+        radius = coreRadius,
+        center = center,
+    )
+}
+
+/**
  * Draw the cytoplasm motes — memory graph nodes as small luminous
  * points drifting between the nucleus and the membrane.
  *
  * Positioning formula per mote (deterministic, no stored state):
- *  - baseAngleDeg = stableIndex × 137.508 (golden angle) mod 360
- *  - baseRadialFrac = sqrt((stableIndex + 0.5) / totalCount)
- *     → uniform 2D density (each equal-area ring has ~equal motes)
- *  - radial = lerp(nucleusRadius × 1.10, membraneRadius × 0.92, frac)
- *  - driftX = sin(t × wX + phaseX) × amp
- *  - driftY = cos(t × wY + phaseY) × amp
+ *   baseAngleDeg  = stableIndex × GOLDEN_ANGLE_DEG mod 360
+ *   baseRadialFrac = sqrt((stableIndex + 0.5) / totalCount)
+ *   radial        = lerp(nucleusRadius × 1.10, membraneRadius × 0.92, frac)
+ *   drift         = per-mote sin/cos using phases derived from index
  *
- * Per-mote frequencies and phases come from the stableIndex so every
- * mote drifts a little differently — the cloud looks alive without
- * any mote mutating its own state.
- *
- * Newly-born motes fade in over [cfg.moteBirthMs] with a brief white
- * halo, then settle into ambient rendering.
+ * Newly-born motes fade in over [CellVizConfig.moteBirthMs] with a
+ * brief white halo, then settle into ambient rendering.
  */
 private fun DrawScope.drawCytoplasmMotes(
     motes: List<CytoplasmMote>,
@@ -1238,7 +597,7 @@ private fun DrawScope.drawCytoplasmMotes(
     motes.forEach { mote ->
         val idx = mote.stableIndex
 
-        // Base position via golden-angle scatter + sqrt radial.
+        // Base position: golden-angle scatter + sqrt radial
         val angleDeg = ((idx.toFloat() * GOLDEN_ANGLE_DEG) % 360f + 360f) % 360f
         val radialFrac = kotlin.math.sqrt((idx + 0.5f) / totalCount.toFloat())
             .coerceIn(0f, 1f)
@@ -1247,9 +606,7 @@ private fun DrawScope.drawCytoplasmMotes(
         val bx = cx + r * cos(angleRad).toFloat()
         val by = cy + r * sin(angleRad).toFloat()
 
-        // Per-mote drift. Derived from index so it's deterministic.
-        // Each mote gets a unique (phase, freq-multiplier) without any
-        // RNG call in the hot path.
+        // Per-mote drift — deterministic from index, no RNG in hot path
         val phaseX = (idx * 0.7531f) % (2f * PI.toFloat())
         val phaseY = (idx * 1.2847f) % (2f * PI.toFloat())
         val wX = driftOmega * (0.85f + 0.30f * ((idx * 31) % 17) / 17f)
@@ -1259,24 +616,21 @@ private fun DrawScope.drawCytoplasmMotes(
 
         val pos = Offset(bx + dx, by + dy)
 
-        // Birth animation — fade-in + brief bright halo.
-        val birthAge = if (mote.birthTimeMs > 0L) nowMs - mote.birthTimeMs else Long.MAX_VALUE
+        // Birth animation
+        val birthAge = if (mote.birthTimeMs > 0L)
+            nowMs - mote.birthTimeMs
+        else Long.MAX_VALUE
         val birthProgress = if (cfg.moteBirthMs > 0L)
             (birthAge.toFloat() / cfg.moteBirthMs).coerceIn(0f, 1f)
         else 1f
-        val birthScale = smoothstep(birthProgress)  // 0 → 1
+        val birthScale = smoothstep(birthProgress)
 
-        // Scope-tinted color from the current theme. getScopeColor picks
-        // the right accent so motes recolor when the user changes theme.
         val moteColor = colorTheme.getScopeColor(mote.scope)
-
-        // Core dot — small, solid-ish. Alpha scaled by birth progress.
         val coreRadius = cfg.moteRadiusPx * birthScale
         if (coreRadius < 0.4f) return@forEach
 
         if (isDarkMode) {
-            // Soft halo in dark mode so motes read like distant stars,
-            // not painted specks. Two stacked circles — no shader.
+            // Two-circle halo so motes read like distant stars.
             drawCircle(
                 color = moteColor.copy(alpha = 0.22f * birthScale),
                 radius = coreRadius * 2.5f,
@@ -1289,12 +643,13 @@ private fun DrawScope.drawCytoplasmMotes(
             )
         }
         drawCircle(
-            color = moteColor.copy(alpha = (if (isDarkMode) 0.95f else 0.70f) * birthScale),
+            color = moteColor.copy(
+                alpha = (if (isDarkMode) 0.95f else 0.70f) * birthScale
+            ),
             radius = coreRadius,
             center = pos,
         )
 
-        // Birth halo — a single white ring pulse over the birth window.
         if (birthProgress < 1f) {
             val haloAlpha = (1f - birthProgress) * 0.85f
             drawCircle(
@@ -1307,11 +662,8 @@ private fun DrawScope.drawCytoplasmMotes(
     }
 }
 
-/** Golden angle in degrees — 137.508° ≈ 360° × (1 − 1/φ) where φ is the golden ratio. */
-private const val GOLDEN_ANGLE_DEG = 137.50776f
-
 // -----------------------------------------------------------------------------
-// Geometry helpers — no Compose dependencies, trivial math
+// Geometry helpers — Compose-dependent
 // -----------------------------------------------------------------------------
 
 /** Convert a polar (radius, degrees) coord to an Offset in screen space. */
@@ -1322,15 +674,13 @@ private fun polar(cx: Float, cy: Float, r: Float, deg: Float): Offset {
 
 /**
  * A diamond aligned so its long axis points radially outward from the
- * cell's center — adapterRotationDeg is the angle from center to the
+ * cell's center. [adapterRotationDeg] is the angle from center to the
  * port's position.
  */
 private fun diamondPath(center: Offset, radius: Float, adapterRotationDeg: Float): Path {
     val rad = adapterRotationDeg.toDouble() * PI / 180.0
     val outward = Offset(cos(rad).toFloat(), sin(rad).toFloat())
     val tangent = Offset(-outward.y, outward.x)
-    // Slight asymmetry — the diamond is a touch longer on its radial
-    // axis than its tangential axis, giving a "docked plug" silhouette.
     val radial  = radius * 1.1f
     val lateral = radius * 0.85f
     return Path().apply {
@@ -1342,9 +692,7 @@ private fun diamondPath(center: Offset, radius: Float, adapterRotationDeg: Float
     }
 }
 
-/**
- * A regular hexagon with one flat oriented tangent to the membrane.
- */
+/** A regular hexagon with one flat oriented tangent to the membrane. */
 private fun hexPath(center: Offset, radius: Float, adapterRotationDeg: Float): Path {
     val path = Path()
     for (i in 0 until 6) {
